Apply to SourceController::read()
