+++ src/Controller/Source/SourceController.php
@@ -6,6 +6,8 @@ namespace App\Controller\Source;
 
 use App\Exception\ServiceException;
 use App\Security\ApiToken;
+use App\ServiceProxy\ServiceProxy;
+use App\ServiceRequest\RequestBuilderFactory;
 use Psr\Http\Client\ClientExceptionInterface;
 use SmartAssert\ServiceClient\Exception\HttpResponseExceptionInterface;
 use SmartAssert\ServiceClient\Exception\InvalidModelDataException;
@@ -13,10 +15,9 @@ use SmartAssert\ServiceClient\Exception\InvalidResponseDataException;
 use SmartAssert\ServiceClient\Exception\InvalidResponseTypeException;
 use SmartAssert\ServiceClient\Exception\UnauthorizedException;
 use SmartAssert\ServiceClient\SerializableInterface;
-use SmartAssert\SourcesClient\FileSourceClientInterface;
-use SmartAssert\SourcesClient\GitSourceClientInterface;
 use SmartAssert\SourcesClient\SourceClientInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -24,8 +25,8 @@ readonly class SourceController
 {
     public function __construct(
         private SourceClientInterface $client,
-        private FileSourceClientInterface $fileSourceClient,
-        private GitSourceClientInterface $gitSourceClient,
+        private RequestBuilderFactory $requestBuilderFactory,
+        private ServiceProxy $sourcesProxy,
     ) {
     }
 
@@ -59,30 +60,21 @@ readonly class SourceController
     }
 
     /**
-     * @param non-empty-string $sourceId
-     *
      * @throws ServiceException
-     * @throws UnauthorizedException
      */
     #[Route(path: '/source/{sourceId<[A-Z90-9]{26}>}', name: 'source_read', methods: ['GET'])]
-    public function read(ApiToken $token, string $sourceId): Response
+    public function read(ApiToken $token, Request $request): Response
     {
-        try {
-            try {
-                $source = $this->gitSourceClient->get($token->token, $sourceId);
-            } catch (InvalidModelDataException) {
-                $source = $this->fileSourceClient->get($token->token, $sourceId);
-            }
+        $requestBuilder = $this->requestBuilderFactory->create('GET', $request->getRequestUri());
+        $httpRequest = $requestBuilder
+            ->withAuthorization($token->token)
+            ->get()
+        ;
 
-            return new JsonResponse($source->toArray());
-        } catch (
-            ClientExceptionInterface |
-            HttpResponseExceptionInterface |
-            InvalidModelDataException |
-            InvalidResponseDataException |
-            InvalidResponseTypeException $e
-        ) {
-            throw new ServiceException('sources', $e);
+        try {
+            return $this->sourcesProxy->sendRequest($httpRequest);
+        } catch (ClientExceptionInterface $exception) {
+            throw new ServiceException('sources', $exception);
         }
     }
 }
