+++ tests/Application/FileSource/AbstractDeleteTest.php
@@ -52,7 +52,7 @@ abstract class AbstractDeleteTest extends AbstractApplicationTestCase
         $id = $this->createFileSource($apiKey->key, $label);
 
         $getResponse = $this->applicationClient->makeGetSourceRequest($apiKey->key, $id);
-        $this->assertRetrievedFileSource($getResponse, $label, $id);
+        $this->assertRetrievedFileSource($getResponse, $label, $user->id, $id);
 
         $deleteResponse = $this->applicationClient->makeDeleteFileSourceRequest($apiKey->key, $id);
         $this->assertDeletedFileSource($deleteResponse, $label, $user->id, $id);
