+++ tests/Functional/Controller/Source/FileSourceControllerTest.php
@@ -5,17 +5,12 @@ declare(strict_types=1);
 namespace App\Tests\Functional\Controller\Source;
 
 use App\Tests\Application\AbstractApplicationTestCase;
-use App\Tests\Exception\Http\ClientException;
+use App\Tests\DataProvider\ServiceExceptionDataProviderTrait;
 use App\Tests\Functional\Controller\AssertJsonResponseTrait;
 use App\Tests\Functional\GetClientAdapterTrait;
 use GuzzleHttp\Handler\MockHandler;
-use GuzzleHttp\Psr7\Response;
-use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Client\ClientInterface as HttpClientInterface;
-use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use SmartAssert\ServiceClient\Exception\CurlException;
-use SmartAssert\ServiceClient\Exception\CurlExceptionInterface;
 use SmartAssert\UsersClient\ClientInterface as UsersClient;
 use SmartAssert\UsersClient\Model\Token;
 use Symfony\Component\Uid\Ulid;
@@ -24,6 +19,7 @@ class FileSourceControllerTest extends AbstractApplicationTestCase
 {
     use GetClientAdapterTrait;
     use AssertJsonResponseTrait;
+    use ServiceExceptionDataProviderTrait;
 
     /**
      * @dataProvider serviceExceptionDataProvider
@@ -172,89 +168,5 @@ class FileSourceControllerTest extends AbstractApplicationTestCase
         $response = $this->applicationClient->makeFileSourceFilesRequest($apiKey, $sourceId);
 
         $this->assertJsonResponse($response, $expectedStatusCode, $expectedData);
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function serviceExceptionDataProvider(): array
-    {
-        $exceptionMessage = md5((string) rand());
-        $exceptionCode = rand();
-        $serviceName = 'sources';
-
-        return [
-            ClientExceptionInterface::class => [
-                'httpFixture' => new ClientException(
-                    $exceptionMessage,
-                    $exceptionCode
-                ),
-                'expectedStatusCode' => 500,
-                'expectedData' => [
-                    'type' => 'service-communication-failure',
-                    'context' => [
-                        'service' => $serviceName,
-                        'error' => [
-                            'code' => $exceptionCode,
-                            'message' => $exceptionMessage,
-                        ],
-                    ],
-                ],
-            ],
-            CurlExceptionInterface::class => [
-                'httpFixture' => new CurlException(
-                    \Mockery::mock(RequestInterface::class),
-                    $exceptionCode,
-                    $exceptionMessage,
-                ),
-                'expectedStatusCode' => 500,
-                'expectedData' => [
-                    'type' => 'service-communication-failure',
-                    'context' => [
-                        'service' => $serviceName,
-                        'error' => [
-                            'code' => $exceptionCode,
-                            'message' => $exceptionMessage,
-                        ],
-                    ],
-                ],
-            ],
-            '405, no response content type' => [
-                'httpFixture' => new Response(
-                    status: 405,
-                    reason: 'Method not allowed.'
-                ),
-                'expectedStatusCode' => 500,
-                'expectedData' => [
-                    'type' => 'service-communication-failure',
-                    'context' => [
-                        'service' => $serviceName,
-                        'code' => 405,
-                        'reason' => 'Method not allowed.',
-                        'expected_content_type' => 'application/json',
-                        'actual_content_type' => null,
-                    ],
-                ],
-            ],
-            '200, text/html content type' => [
-                'httpFixture' => new Response(
-                    status: 200,
-                    headers: ['content-type' => 'text/html'],
-                    body: '<html />',
-                    reason: 'Ok.'
-                ),
-                'expectedStatusCode' => 500,
-                'expectedData' => [
-                    'type' => 'service-communication-failure',
-                    'context' => [
-                        'service' => $serviceName,
-                        'code' => 200,
-                        'reason' => 'Ok.',
-                        'expected_content_type' => 'application/json',
-                        'actual_content_type' => 'text/html',
-                    ],
-                ],
-            ],
-        ];
     }
 }
